Full code audit and optimization of SystemMonitor
This commit implements several fixes and optimizations identified during
a full code audit:

- Optimized ProcessView::UpdateThread to skip thread scanning for idle
  teams (zero CPU delta), reducing system call overhead.
- Enhanced Utils::FormatBytes to support TiB and PiB units.
- Standardized Utils::FormatHertz to use 'kHz' (lowercase k).
- Implemented time-limited caching (5-10s TTL) for frequently accessed
  system info getters (Display, Disk, IP, Battery, etc.) in Utils.cpp.
- Refined ActivityGraphView rendering to include an extra data point at
  the rightmost edge, ensuring graphs cover the full view width.
- Aligned System Information reporting between SystemSummaryView and
  SystemDetailsView (consistent RAM/Swap labels and formatting).
- Standardized CircularBuffer.h and DataHistory.cpp to use uint32 for
  all size and index-related variables for better type safety.
- Verified all core data structure changes with unit tests.

+++ SystemMonitor/DataHistory.cpp
@@ -57,7 +57,7 @@ DataHistory::ValueAt(bigtime_t time, int32* hintIndex)
 	if (hintIndex != NULL && *hintIndex >= 0)
 		left = *hintIndex;
 
-	int32 right = fBuffer.CountItems() - 1;
+	int32 right = (int32)fBuffer.CountItems() - 1;
 	data_item* item = NULL;
 
 	while (left <= right) {
@@ -161,8 +161,8 @@ DataHistory::_ResetDeques()
 	fMinDeque.clear();
 	fMaxDeque.clear();
 
-	int32 count = fBuffer.CountItems();
-	for (int32 i = 0; i < count; i++) {
+	uint32 count = fBuffer.CountItems();
+	for (uint32 i = 0; i < count; i++) {
 		data_item* item = fBuffer.ItemAt(i);
 		if (item == NULL) continue;
